+++ lib/screen/home_screen.dart
@@ -1,6 +1,9 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:intl/intl.dart';
+import 'package:monitor_app/constants/constants.dart';
+import 'package:monitor_app/controller/app_provider.dart';
 
 import 'package:monitor_app/screen/components/task_card.dart';
 import 'package:monitor_app/constants/strings.dart';
@@ -21,21 +24,50 @@ class HomeScreen extends ConsumerStatefulWidget {
 }
 
 class HomeScreenState extends ConsumerState<HomeScreen> {
+  late DateTimeRange dateRange;
+
   @override
   void initState() {
     super.initState();
+    dateRange = DateTimeRange(
+      start: DateTime.now().subtract(const Duration(days: 30)),
+      end: DateTime.now(),
+    );
     WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
       debugPrint('HomeScreen initState : email ${widget.email}');
-      ref.read(taskControllerProvider.notifier).getAllTasks(widget.email);
+      ref.read(taskControllerProvider.notifier).getAllTasks(
+            email: widget.email,
+            dateTimeRange: dateRange,
+          );
     });
   }
 
   @override
   Widget build(BuildContext context) {
+    final String status = ref.watch(statusTaskProvider);
+    ref.listen<TaskState>(taskControllerProvider,
+        (TaskState? prevState, TaskState newState) async {
+      if (newState is PrepareTaskSuccess) {
+        await Navigator.of(context)
+            .pushNamed(TaskScreen.routeName, arguments: newState.task)
+            .then((_) => Future(
+                () => ref.read(taskControllerProvider.notifier).getAllTasks(
+                      email: widget.email,
+                      dateTimeRange: dateRange,
+                      status: status,
+                    )));
+      }
+    });
     return Scaffold(
       appBar: AppBar(
         forceMaterialTransparency: false,
         title: const Text(titleHomeScreen),
+        actions: [
+          IconButton(
+            onPressed: () async => _buildModalBottomSheet(),
+            icon: const Icon(Icons.filter_alt_rounded),
+          ),
+        ],
       ),
       body: _getBody(),
     );
@@ -51,6 +83,7 @@ class HomeScreenState extends ConsumerState<HomeScreen> {
 
   Widget _getBody() {
     final state = ref.watch(taskControllerProvider);
+    final String status = ref.watch(statusTaskProvider);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
@@ -62,7 +95,7 @@ class HomeScreenState extends ConsumerState<HomeScreen> {
             child: Consumer(
               builder: (context, ref, child) {
                 // debugPrint(state.toString());
-                if (state is TasksLoaded) {
+                if (state is TaskLoaded) {
                   if (state.tasks.isEmpty) {
                     return const Center(child: Text('No Task'));
                   } else {
@@ -71,12 +104,21 @@ class HomeScreenState extends ConsumerState<HomeScreen> {
                         return TaskCard(
                           task: state.tasks[index],
                           onSelectTask: () async {
-                            await Navigator.of(context)
-                                .pushNamed(TaskScreen.routeName,
-                                    arguments: state.tasks[index])
-                                .then((_) => Future(() => ref
-                                    .read(taskControllerProvider.notifier)
-                                    .getAllTasks(widget.email)));
+                            if (state.tasks[index].status == STATUS_TODO) {
+                              ref
+                                  .read(taskControllerProvider.notifier)
+                                  .isTaskFoundInLocal(task: state.tasks[index]);
+                            } else {
+                              await Navigator.of(context)
+                                  .pushNamed(TaskScreen.routeName,
+                                      arguments: state.tasks[index])
+                                  .then((_) => Future(() => ref
+                                      .read(taskControllerProvider.notifier)
+                                      .getAllTasks(
+                                          email: widget.email,
+                                          dateTimeRange: dateRange,
+                                          status: status)));
+                            }
                           },
                         );
                       },
@@ -98,55 +140,182 @@ class HomeScreenState extends ConsumerState<HomeScreen> {
     );
   }
 
-  Container _buildSortAndFilter() {
-    return Container(
-      color: Colors.white,
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          ElevatedButton.icon(
-            style: ElevatedButton.styleFrom(
-              padding: const EdgeInsets.symmetric(horizontal: 24),
-              side: const BorderSide(color: Color.fromARGB(255, 37, 33, 33)),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20)),
-              foregroundColor: Colors.white,
-              elevation: 0,
-              shadowColor: Colors.transparent,
+  // Container _buildSortAndFilter() {
+  //   return Container(
+  //     color: Colors.white,
+  //     padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+  //     child: Row(
+  //       mainAxisAlignment: MainAxisAlignment.spaceBetween,
+  //       children: [
+  //         ElevatedButton.icon(
+  //           style: ElevatedButton.styleFrom(
+  //             padding: const EdgeInsets.symmetric(horizontal: 24),
+  //             side: const BorderSide(color: Color.fromARGB(255, 37, 33, 33)),
+  //             shape: RoundedRectangleBorder(
+  //                 borderRadius: BorderRadius.circular(20)),
+  //             foregroundColor: Colors.white,
+  //             elevation: 0,
+  //             shadowColor: Colors.transparent,
+  //           ),
+  //           onPressed: () {
+  //             _buildModalBottomSheet();
+  //           },
+  //           icon: const Icon(
+  //             Icons.filter_alt_rounded,
+  //             color: Colors.black,
+  //           ),
+  //           label: const Text(
+  //             'Filter',
+  //             style: TextStyle(color: Colors.black),
+  //           ),
+  //         ),
+  //         // ElevatedButton.icon(
+  //         //   style: ElevatedButton.styleFrom(
+  //         //     padding: const EdgeInsets.symmetric(horizontal: 24),
+  //         //     side: const BorderSide(color: Color.fromARGB(255, 37, 33, 33)),
+  //         //     shape: RoundedRectangleBorder(
+  //         //         borderRadius: BorderRadius.circular(20)),
+  //         //     foregroundColor: Colors.white,
+  //         //     elevation: 0,
+  //         //     shadowColor: Colors.transparent,
+  //         //   ),
+  //         //   onPressed: () {},
+  //         //   icon: const Icon(
+  //         //     Icons.filter_alt_outlined,
+  //         //     color: Colors.black,
+  //         //   ),
+  //         //   label: const Text(
+  //         //     'fILTER',
+  //         //     style: TextStyle(color: Colors.black),
+  //         //   ),
+  //         // ),
+  //       ],
+  //     ),
+  //   );
+  // }
+
+  Future<dynamic> _buildModalBottomSheet() {
+    return showModalBottomSheet(
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(20),
+          topRight: Radius.circular(20),
+        ),
+      ),
+      backgroundColor: Colors.white,
+      context: context,
+      builder: (context) => StatefulBuilder(
+        builder: (context, StateSetter setState) {
+          return Container(
+            padding: const EdgeInsets.all(20),
+            height: 300,
+            width: double.infinity,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const Text('Range Date :'),
+                ElevatedButton.icon(
+                  style: ElevatedButton.styleFrom(
+                    padding: const EdgeInsets.symmetric(horizontal: 24),
+                    side: const BorderSide(
+                        color: Color.fromARGB(255, 187, 201, 230)),
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(10)),
+                    foregroundColor: Colors.white,
+                    elevation: 0,
+                    shadowColor: Colors.transparent,
+                  ),
+                  onPressed: () {
+                    showDateRangePicker(
+                            context: context,
+                            firstDate: DateTime(2022),
+                            lastDate: DateTime(2040),
+                            initialDateRange: dateRange,
+                            initialEntryMode: DatePickerEntryMode.calendarOnly)
+                        .then((value) {
+                      setState(() {
+                        if (value != null) dateRange = value;
+                        debugPrint(
+                            'periode : ${DateFormat('dd/MM/yyyy').format(dateRange.start)}  -  ${DateFormat('dd/MM/yyyy').format(dateRange.end)}');
+                      });
+                    });
+                  },
+                  icon: const Icon(
+                    Icons.calendar_month_rounded,
+                    color: Colors.black,
+                  ),
+                  label: Text(
+                    '${DateFormat('dd/MM/yyyy').format(dateRange.start)}  -  ${DateFormat('dd/MM/yyyy').format(dateRange.end)} ',
+                    style: const TextStyle(color: Colors.black),
+                  ),
+                ),
+                const SizedBox(height: 20),
+                const Text('Status Task :'),
+                _buildDropdownStatus(),
+                const Spacer(),
+                Consumer(builder: (context, ref, child) {
+                  final String status = ref.watch(statusTaskProvider);
+                  return InkWell(
+                    onTap: () {
+                      ref.read(taskControllerProvider.notifier).getAllTasks(
+                            email: widget.email,
+                            dateTimeRange: dateRange,
+                            status: status,
+                          );
+                    },
+                    child: Container(
+                      width: MediaQuery.of(context).size.width,
+                      padding: const EdgeInsets.all(10),
+                      decoration: BoxDecoration(
+                        color: Colors.blue,
+                        borderRadius: BorderRadius.circular(10),
+                      ),
+                      child: const Center(
+                        child: Text('Apply',
+                            style: TextStyle(
+                                color: Colors.white,
+                                fontWeight: FontWeight.bold)),
+                      ),
+                    ),
+                  );
+                }),
+              ],
             ),
-            onPressed: () {},
-            icon: const Icon(
-              Icons.arrow_upward,
-              color: Colors.black,
-            ),
-            label: const Text(
-              'Terbaru',
-              style: TextStyle(color: Colors.black),
-            ),
-          ),
-          ElevatedButton.icon(
-            style: ElevatedButton.styleFrom(
-              padding: const EdgeInsets.symmetric(horizontal: 24),
-              side: const BorderSide(color: Color.fromARGB(255, 37, 33, 33)),
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20)),
-              foregroundColor: Colors.white,
-              elevation: 0,
-              shadowColor: Colors.transparent,
-            ),
-            onPressed: () {},
-            icon: const Icon(
-              Icons.filter_alt_outlined,
-              color: Colors.black,
-            ),
-            label: const Text(
-              'fILTER',
-              style: TextStyle(color: Colors.black),
-            ),
-          ),
-        ],
+          );
+        },
       ),
     );
+  }
+
+  List<String> dataStatus = ['All', 'Todo', 'Review', 'Verified'];
+  Widget _buildDropdownStatus() {
+    return Consumer(builder: (_, WidgetRef ref, __) {
+      var status = ref.watch(statusTaskProvider);
+      debugPrint('status : $status');
+      return Container(
+        padding: const EdgeInsets.symmetric(horizontal: 30),
+        decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(10),
+            border: Border.all(width: 1, color: const Color.fromARGB(255, 187, 201, 230))),
+        child: DropdownButton(
+          underline: const SizedBox(),
+          borderRadius: BorderRadius.circular(10),
+          value: status.isNotEmpty ? status : null,
+          onChanged: (value) {
+            debugPrint('value : $value');
+            // ref.read(taskControllerProvider.notifier).getAllTasks(
+            //     email: widget.email, dateTimeRange: dateRange, status: value);
+            ref.read(statusTaskProvider.notifier).state = value!;
+          },
+          items: dataStatus.map<DropdownMenuItem<String>>((String value) {
+            return DropdownMenuItem<String>(
+              value: value,
+              child: Text(value),
+            );
+          }).toList(),
+        ),
+      );
+    });
   }
 }