+++ lib/screen/task_screen.dart
@@ -131,7 +131,7 @@ class _TaskScreenState extends ConsumerState<TaskScreen> {
                     width: 0,
                   ),
             Visibility(
-              visible: false,
+              visible: true,
               child: IconButton(
                 onPressed: () async {
                   // progressDialogue();
@@ -140,13 +140,13 @@ class _TaskScreenState extends ConsumerState<TaskScreen> {
                   for (var asset in assets!) {
                     if (idx % 3 == 0) {
                       asset.url =
-                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699513188710.jpg";
+                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699843714069.jpg";
                     } else if (idx % 3 == 1) {
                       asset.url =
-                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699513191279.jpg";
+                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699843719016.jpg";
                     } else {
                       asset.url =
-                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699513188710.jpg";
+                          "/storage/emulated/0/Android/data/com.bci.monitor_app/files/1699843714069.jpg";
                     }
                     idx++;
                     await ref
@@ -162,7 +162,7 @@ class _TaskScreenState extends ConsumerState<TaskScreen> {
         body: widget.task.status == STATUS_TODO
             ? Consumer(builder: (context, ref, child) {
                 debugPrint('reload task screen $state');
-                if (state is TasksLoaded) {
+                if (state is TaskLoaded) {
                   task = state.tasks[0];
                   return Column(
                     children: [
@@ -247,7 +247,7 @@ class _TaskScreenState extends ConsumerState<TaskScreen> {
                   SvgPicture.asset('assets/icons/ic_report_torque.svg',
                       width: 40, height: 40),
                   const SizedBox(width: 20),
-                  const Text('TORQUE BOLT TIGHTENING',
+                  const Text('BOLT TIGHTENING TORQUE',
                       style: TextStyle(
                           color: Colors.white, fontWeight: FontWeight.bold)),
                 ],